Simplify determining which nodes are "alien".
Remove the hacky "motionNode" field from Plan struct. It was set to the
the Motion node at the root of each slice, but we have a better mechanism
to track the slice that each executor node belongs to. Namely, there's
a field, currentSliceIdInPlan, in EState, which is kept set to the current
slice as we traverse the executor tree in ExecInitPlan().