Stop using Flows in executor, except during initialization.
Flow was used to get the number of receiving segments for the hash
calculation in in Motion and Result nodes. Use the number of segments
from the slice table instead. Seems more clear, and removes some ugly
differences in handling Postgres planner vs. ORCA plans.

In an upcoming commit, I will move the responsibility of generating the
slice table from the executor to the planner. Once that's done, there will
no more references to Flows in the executor at all. For now, though, we're
still using the flows to build the slice table at executor startup.

Reviewed-by: Taylor Vesely <tvesely@pivotal.io>