Attach InitPlans to Motion nodes, like any other node.
Fix the code in cdbllize.c that stripped off nodes from a Plan tree to
not lose track of InitPlans attached to nodes that are stripped off. I
think that was always wrong, but it wasn't exposed before, when we didn't
attach InitPlans to Motions. But it was an accident waiting to happen.
Without fixing that, there was a regression failure in 'qp_subquery' test.

Per Taylor's question in PR comment.