Dumb down the HINT when you try to drop/rename a partition directly.
Why is it a good idea to purposefully make a HINT less precise, you might
ask? This is part of a series of commits that take us a little bit closer
to the upstream partitioning implementation. PostgreSQL doesn't allow
the FOR RANK() way of addressing partitions, and I'm about to remove that
in the next commit. Without FOR RANK(), there are some corner-cases where
there is no way to address a partition, so we don't want to give a HINT
that suggests doing that. But in PostgreSQL, it is possible to use DROP
TABLE or ALTER TABLE RENAME directly on a partition, so we won't need to
give these errors anymore.

It's not much, but this reduces the difference between Greenplum and
PostgreSQL a little.