Store containing slice's ID in every Plan node.
This gets rid of the step in executor startup that walked through the plan
tree, and labeled every plan node with the parent Motion node. We can do
that conveniently at the same time that we build the slice table. Modifying
the motionNode field in the Plan tree in an already-built plan tree
always felt iffy.