Remove support for addressing partitions with FOR(RANK(...))
It doesn't seem like a very useful syntax for users, but we had used it
quite a lot in regression tests. Modify most tests to use the FOR(<value>)
syntax. Leave one in 'partition' test, so that we have some coverage for
the error message you now get with FOR(RANK(...)).

This leaves one unpleasant corner-case:

    create table foopart (i int4) partition by range (i)
      (partition one start(1) inclusive end(2) inclusive)

    alter table foopart add partition start (5) exclusive end (6) exclusive;

The second partition cannot be addressed by name, because it has no name.
But it also cannot be addressed by value, because no value falls into the
range. As things stand, there's no way to drop it. Accept that for now,
because things will change as we merge with PostgreSQL v10 and get the
upstream partitioning implementation. PostgreSQL allows DROP TABLE and
ALTER TABLE RENAME directly on a partition relation, so you can once we
merge with v10, you can do that. It's hopefully not a situation that
people will run into too often.

Discussion: https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/Pag9a9DFpOk/I_xO5pZLBAAJ