Build GPORCA unit tests without CMake.
Build the 'gporca_test' program using same build infrastructure that's
used to build GPORCA as part of the 'postgres' binary. That way, the tests
are built using the same compiler options as when the code is run for
real, and you don't need to build all of GPORCA twice, which should speed
up the overall time to compile Greenplum and run all the unit tests.

The GPORCA unit tests are now run with the same "make unittest-check"
Makefile target as all the other unit tests in the GPDB repository.

To run the tests, we now use the TAP 'prove' program, with a little
perl wrapper. The 'gporca_test' program is modified to produce TAP
compatible output, with a "ok" or "not ok" line.

Co-authored-by: Chris Hajas <chajas@vmware.com>