Remove 'onerow' field, it's not really needed.
The 'onerow' field was used in create_nestloop_path(), to skip creating a
Materialize node on the inner side, if the outer path was guaranteed to
return only one row. However, the similar logic in create_nestloop_plan()
didn't pay attention to the 'onerow' field, and went ahead and created the
Materialize node, anyway.

A second place where 'onerow' was used, was in setting the row count
estimates for subqueries and table functions. That seems redundant: if the
cost estimator can detect the single-row case, it already sets the
estimate for the base relation to 1 row, and that should be propagated up
to the subquery's cost estimate as usual.

I went down this rabbit hole, by looking at the 'num_leading_eq' variable
first. It was used to detect the case that a unique index guarantees that
the relation can return only a single row. It treated "x IS NULL" the same
as "x = <const>", but that is not correct: a unique index can contain
multiple NULL values. So if the query contained an indexable IS NULL qual,
we might incorrectly set the 'onerow' flag. I constructed this test case
for it:

create table foo (id int4);
create table bar (id int4, x int4, unique (id, x));
insert into foo select g from generate_series(1, 100000) g;
insert into bar values (null, 1), (null, 2);
set enable_hashjoin=off;
set enable_nestloop=on;
set enable_seqscan=off;
explain (costs off) select * from bar left join foo on foo.id = bar.x
where bar.id is null and bar.x=1;

I verified with a debugger that 'onerow' is set on the 'bar' relation,
and we incorrectly skip inserting the Materialize node on the inner side
in create_nestloop_path(). However, as you can see from the plan below,
you still get a Materialize node there, because it's still added later, in
create_nestloop_plan():

                          QUERY PLAN
---------------------------------------------------------------
 Gather Motion 3:1  (slice2; segments: 3)
   ->  Nested Loop Left Join
         Join Filter: (foo.id = bar.x)
         ->  Index Only Scan using bar_id_x_key on bar
               Index Cond: ((id IS NULL) AND (x = 1))
         ->  Materialize
               ->  Broadcast Motion 1:3  (slice1; segments: 1)
                     ->  Seq Scan on foo
                           Filter: (id = 1)
 Optimizer: legacy query optimizer
(10 rows)

Given that we are inserting the Materialize node in
create_nestloop_plan() stage anyway, it's pointless to try avoiding it
in the create_nestloop_path() stage. I considered ripping out the
code in create_nestloop_path() altogether, and relying on
create_nestloop_plan() to do it in all cases. But it seems marginally
better to include it in the Path already, because otherwise the cost of
the extra Materialize node is not included in the Path's cost estimate.

As a concept, it's not a bad idea to prove that a relation can only
return single row, and use that fact in optimization. There is some code
like that in newer versions of PostgreSQL, to try to prove a relation as
unique, and use that to remove unnecessary left joins, and to reduce
semijoins to inner joins.

Reviewed-by: Paul Guo <pguo@pivotal.io>